requirements.txt
apify-client>=1.6.0
python-dotenv>=1.0.0
openai>=1.0.0
